back-end

+++ src/controllers/videoController.js
@@ -5,6 +5,7 @@ export const home = async(req, res) => {
 	const videos = await Video.find({}).sort({createdAt:"desc"});
 	return res.render("home", {pageTitle: "Home", videos});
 }
+
 export const watch = async (req, res) => {
 	const { id } = req.params;
 	const video = await Video.findById(id).populate("owner"); // populate은 owner부분을 실제 user객체로 채워준다.
@@ -14,9 +15,13 @@ export const watch = async (req, res) => {
 }
 export const getEdit = async (req, res) => {
 	const { id } = req.params;
+	const {user: {_id}} = req.session;
 	const video = await Video.findById(id);
 	if (!video)
 		return res.status(404).render("404", {pageTitle: "Video not found."});
+	if (String(video.owner) !== String(_id)){
+		return res.status(403).redirect("/");
+	}
 	return res.render("edit", {pageTitle: `Editing ${video.title}`, video});
 }
 export const postEdit = async (req, res) => {
@@ -58,13 +63,16 @@ export const postUpload = async (req, res) => {
 	const { path: fileUrl } = req.file;
 	const { title, description, hashtags } = req.body;
 	try {
-		await Video.create({
+		const newVideo = await Video.create({
 			title,
 			fileUrl: fileUrl,
 			description,
 			hashtags: Video.formatHashtags(hashtags),
 			owner: _id,
 		});
+		const user = await User.findById(_id);
+		user.videos.push(newVideo._id);
+		user.save();
 		return res.redirect("/");
 	 } catch(error){
 		console.log(error);
@@ -76,7 +84,15 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async(req, res) => {
+	const {user: {_id}} = req.session;
 	const { id } = req.params;
+	const video = await Video.findById(id);
+	if (!video){
+		return res.status(404).render("404", {pageTitle: "Video not found."});
+	}
+	if (String(video.owner) !== String(_id)){
+		return res.status(403).redirect("/");
+	}
 	await Video.findByIdAndDelete(id);
-	return res.send("Delete Video");
+	return res.redirect("/");
 };
